refactor(web): split the 1349-line LineGraph.vue into focused units
LineGraph.vue had grown to 1349 lines — ~4x the next-largest component —
mixing reactive chart geometry, a large SVG template, and ~490 lines of
scoped styles in one file. The business logic was already in composables;
the bloat was geometry plumbing + CSS.

Extract two cohesive units, no behavior change:
- useChartGeometry.js — all pure reactive geometry (domains, scales, the
  line/projection paths, and the pace sub-graph data), matching the
  existing composable pattern (usePanZoom, usePointsData, etc.).
- LineGraph.css — the scoped styles, linked via `<style scoped src>` so
  scoping is preserved.

LineGraph.vue drops 1349 -> 685 lines and is now orchestration + template.
Dead chart-util / rankColors imports removed.

Verified: eslint clean, 75/75 unit tests pass, production build succeeds
(scoped CSS emitted). Audited that no internal-only geometry binding leaks
into the template and every re-exported value is still used.
