+++ packages/web/src/components/LineGraph.vue
@@ -380,27 +380,12 @@ import {
   formatDate,
 } from "@/utils/date";
 import { buildCSV } from "@/utils/csv";
-import {
-  MS_PER_DAY,
-  calcXDomain,
-  calcYDomain,
-  calcPaceYDomain,
-  scaleXFactory,
-  scaleYFactory,
-  buildPathD,
-  buildXTicks,
-  buildAveragePacePath,
-  buildDeviationWedgePath,
-  buildRequiredPaceData,
-  buildPointsEarnedData,
-  requiredPerDayFromBaseline,
-} from "@/utils/chart";
 import { loadSeasonJson } from "@/utils/season";
-import { buildRankBands } from "@/utils/rankColors";
 import { useRankInfo } from "@/composables/useRankInfo";
 import { usePanZoom } from "@/composables/usePanZoom";
 import { useGraphSettings } from "@/composables/useGraphSettings";
 import { usePointsData } from "@/composables/usePointsData";
+import { useChartGeometry } from "@/composables/useChartGeometry";
 import GoalControls from "@/components/GoalControls.vue";
 import StatsPanel from "@/components/StatsPanel.vue";
 import ImportModal from "@/components/modals/ImportModal.vue";
@@ -496,89 +481,43 @@ const {
 const showImportModal = ref(false);
 const showPointsModal = ref(false);
 
-const xDomain = computed(() =>
-  calcXDomain(seasonStart.value, seasonEnd.value, today)
-);
-
-// Filter raw points to the current season (x) domain for display/scaling
-const pointsInSeason = computed(() => {
-  const [min, max] = xDomain.value;
-  return points.filter((p) => {
-    const ms = dateToMs(p.date);
-    return isFinite(ms) && ms >= min && ms <= max;
-  });
-});
-
-const yDomain = computed(() =>
-  calcYDomain(pointsInSeason.value, goalWinPoints.value)
-);
-
-// Scales
-const scaleX = (dateStr) =>
-  scaleXFactory(xDomain.value, width, padding)(dateStr);
-const scaleY = (y) => scaleYFactory(yDomain.value, height, padding)(y);
-
-// Rank overlay bands
-const rankBands = computed(() => buildRankBands(props.goalOptions));
-
-// Points within season, but sorted by date for path rendering
-const sortedPointsInSeason = computed(() => {
-  const [min, max] = xDomain.value;
-  return sortedPoints.value.filter((p) => {
-    const ms = dateToMs(p.date);
-    return ms >= min && ms <= max;
-  });
-});
-
-const scaledPoints = computed(() =>
-  sortedPointsInSeason.value.map((p) => ({ x: scaleX(p.date), y: scaleY(p.y) }))
-);
-
-// Anchor that pace/projection math is measured from. World Tour starts every
-// season at 0, so the anchor is (seasonStart, 0). Ranked starts at a placement
-// rank, so the anchor is the first recorded point — pace, projections and
-// "points earned" then measure progress *since placement*, not from 0.
-const paceBaseline = computed(() => {
-  const sp = sortedPointsInSeason.value;
-  if (props.mode === "ranked" && sp.length) {
-    return { ms: dateToMs(sp[0].date), dateStr: sp[0].date, y: Number(sp[0].y) || 0 };
-  }
-  return { ms: xDomain.value[0], dateStr: seasonStart.value, y: 0 };
-});
-
-// Build path points with a synthetic baseline at season start (y=0) when the
-// first in-season point occurs after the season start. This extends the blue
-// line visually back to day zero without adding a visible point. Ranked is
-// exempt: its line starts at the placement point, with the pre-placement span
-// drawn as a separate flat baseline (see placementBaselinePath).
-const scaledPathPoints = computed(() => {
-  const sp = scaledPoints.value;
-  if (!sp.length) return sp;
-  if (props.mode === "ranked") return sp;
-  const seasonStartMs = dateToMs(seasonStart.value);
-  const firstMs = dateToMs(sortedPointsInSeason.value[0].date);
-  if (isFinite(seasonStartMs) && isFinite(firstMs) && firstMs > seasonStartMs) {
-    const baseline = { x: scaleX(seasonStart.value), y: scaleY(0) };
-    return [baseline, ...sp];
-  }
-  return sp;
-});
-
-const pathD = computed(() => buildPathD(scaledPathPoints.value));
-
-// Ranked only: a flat line from season start across to the placement point at
-// the placement RS, making it clear the climb begins at placement (not 0).
-const placementBaselinePath = computed(() => {
-  if (props.mode !== "ranked") return "";
-  const sp = sortedPointsInSeason.value;
-  if (!sp.length) return "";
-  const seasonStartMs = dateToMs(seasonStart.value);
-  const firstMs = dateToMs(sp[0].date);
-  if (!(isFinite(seasonStartMs) && isFinite(firstMs) && firstMs > seasonStartMs)) return "";
-  const x1 = scaleX(seasonStart.value);
-  const x2 = scaleX(sp[0].date);
-  const y = scaleY(Number(sp[0].y) || 0);
-  return `M${x1},${y} L${x2},${y}`;
+// Chart geometry (domains, scales, line/projection paths, pace sub-graph data).
+// Pure reactive derivations — see useChartGeometry.
+const {
+  scaleY,
+  rankBands,
+  scaledPoints,
+  pathD,
+  placementBaselinePath,
+  pathGoalFromZero,
+  pathGoalFromLast,
+  pathAveragePace,
+  pathDeviationWedge,
+  xTicks,
+  paceScaleY,
+  scaledPaceRequired,
+  scaledPaceEarned,
+  paceRequiredPath,
+  paceEarnedPath,
+  requiredPerDayZero,
+  isFromLastDefined,
+  requiredPerDayFromLast,
+} = useChartGeometry({
+  width,
+  height,
+  padding,
+  paceTop,
+  paceHeight,
+  pacePadY,
+  today,
+  mode: props.mode,
+  goalOptions: () => props.goalOptions,
+  seasonStart,
+  seasonEnd,
+  goalWinPoints,
+  isSeasonValid,
+  points,
+  sortedPoints,
 });
 
 // Rank info based on goalOptions thresholds
@@ -596,119 +535,6 @@ const {
   onPointerMove,
   onPointerUp,
 } = usePanZoom({ width, height, padding }, enableNavigation, navSensitivity);
-
-// Projection paths
-const pathGoalFromZero = computed(() => {
-  if (!isSeasonValid.value) return "";
-  const x1 = scaleX(paceBaseline.value.dateStr);
-  const y1 = scaleY(paceBaseline.value.y);
-  const x2 = scaleX(seasonEnd.value);
-  const y2 = scaleY(goalWinPoints.value);
-  return `M${x1},${y1} L${x2},${y2}`;
-});
-
-const pathGoalFromLast = computed(() => {
-  if (!isSeasonValid.value || sortedPointsInSeason.value.length === 0)
-    return "";
-  const last =
-    sortedPointsInSeason.value[sortedPointsInSeason.value.length - 1];
-  const lastMs = dateToMs(last.date);
-  if (lastMs >= xDomain.value[1]) return "";
-  const x1 = scaleX(last.date);
-  const y1 = scaleY(last.y);
-  const x2 = scaleX(seasonEnd.value);
-  const y2 = scaleY(goalWinPoints.value);
-  return `M${x1},${y1} L${x2},${y2}`;
-});
-
-const pathAveragePace = computed(() => {
-  if (!isSeasonValid.value) return "";
-  return buildAveragePacePath(
-    sortedPointsInSeason.value,
-    xDomain.value[0],
-    xDomain.value[1],
-    scaleX,
-    scaleY,
-    paceBaseline.value
-  );
-});
-
-const pathDeviationWedge = computed(() => {
-  if (!isSeasonValid.value) return "";
-  return buildDeviationWedgePath(
-    sortedPointsInSeason.value,
-    xDomain.value[0],
-    xDomain.value[1],
-    scaleX,
-    scaleY,
-    paceBaseline.value
-  );
-});
-
-const xTicks = computed(() => buildXTicks(xDomain.value, width, padding, 4));
-
-// Pace graph data
-const paceRequiredData = computed(() => {
-  if (!isSeasonValid.value) return [];
-  return buildRequiredPaceData(sortedPointsInSeason.value, goalWinPoints.value, xDomain.value[1]);
-});
-
-const paceEarnedData = computed(() =>
-  buildPointsEarnedData(sortedPointsInSeason.value, paceBaseline.value.y)
-);
-
-const paceYDomain = computed(() =>
-  calcPaceYDomain([
-    ...paceRequiredData.value.map(p => p.y),
-    ...paceEarnedData.value.map(p => p.y),
-  ])
-);
-
-const paceScaleY = (y) => {
-  const [min, max] = paceYDomain.value;
-  return (paceTop + pacePadY) + (1 - (y - min) / (max - min)) * (paceHeight - pacePadY * 2);
-};
-
-const scaledPaceRequired = computed(() =>
-  paceRequiredData.value.map(p => ({ x: scaleX(p.date), y: paceScaleY(p.y) }))
-);
-
-const scaledPaceEarned = computed(() =>
-  paceEarnedData.value.map(p => ({ x: scaleX(p.date), y: paceScaleY(p.y) }))
-);
-
-const paceRequiredPath = computed(() => buildPathD(scaledPaceRequired.value));
-const paceEarnedPath = computed(() => buildPathD(scaledPaceEarned.value));
-
-// Pace stats
-const requiredPerDayZero = computed(() => {
-  // Slope of the baseline→goal projection per day, matching pathGoalFromZero.
-  // World Tour anchors at (seasonStart, 0) → goal over the whole season. Ranked
-  // anchors at the placement point → (goal - placement) over days since placement.
-  const { ms, y } = paceBaseline.value;
-  return requiredPerDayFromBaseline(goalWinPoints.value, y, ms, xDomain.value[1]);
-});
-
-const isFromLastDefined = computed(
-  () =>
-    sortedPointsInSeason.value.length > 0 &&
-    dateToMs(
-      sortedPointsInSeason.value[sortedPointsInSeason.value.length - 1].date
-    ) < xDomain.value[1]
-);
-
-const requiredPerDayFromLast = computed(() => {
-  if (!isFromLastDefined.value) return 0;
-  const last =
-    sortedPointsInSeason.value[sortedPointsInSeason.value.length - 1];
-  const remaining = goalWinPoints.value - last.y;
-  const endMs = xDomain.value[1];
-  const left = Math.max(
-    1,
-    Math.round((endMs - dateToMs(last.date)) / MS_PER_DAY)
-  );
-  return remaining / left;
-});
 
 // Actions
 function setGoalWinPoints(val) {
@@ -856,494 +682,4 @@ try {
 
 </script>
 
-<style scoped>
-.line-graph {
-  max-width: 1000px;
-  margin: 0 auto;
-  text-align: left;
-}
-
-.lg-header {
-  margin-bottom: 8px;
-}
-
-/* Unified status block: goal/progress, today + live, required pace */
-.lg-status {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  margin-bottom: 12px;
-}
-
-/* Overlay pinned across the top of the graph: today's gain on the left,
-   live DB status on the right. pointer-events stay off the container so chart
-   pan/zoom gestures pass through, but the badges themselves keep their tooltip. */
-.graph-overlay {
-  position: absolute;
-  top: 8px;
-  left: 8px;
-  right: 8px;
-  z-index: 10;
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 12px;
-  pointer-events: none;
-}
-.graph-overlay .today-progress,
-.graph-overlay .live-indicator {
-  pointer-events: auto;
-  background: color-mix(in oklab, var(--surface) 78%, transparent);
-  border: 1px solid color-mix(in oklab, var(--primary) 14%, var(--surface));
-  border-radius: 8px;
-  backdrop-filter: blur(2px);
-}
-
-.live-indicator {
-  display: inline-flex;
-  align-items: center;
-  padding: 6px;
-}
-.live-db {
-  display: block;
-}
-/* Connected: steady green database. */
-.live-connected {
-  color: #22c55e;
-}
-/* Connecting: amber database, fading in and out. */
-.live-connecting {
-  color: #eab308;
-}
-@media (prefers-reduced-motion: no-preference) {
-  .live-connecting .live-db {
-    animation: live-fade 1.2s ease-in-out infinite;
-  }
-}
-@keyframes live-fade {
-  0%,
-  100% {
-    opacity: 0.25;
-  }
-  50% {
-    opacity: 1;
-  }
-}
-.lg-gamemode {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 700;
-  letter-spacing: 0.02em;
-  text-transform: uppercase;
-  color: var(--text-strong);
-}
-.lg-season-title {
-  margin: 0;
-  font-size: 22px;
-  font-weight: 900;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  color: var(--text-strong);
-}
-.lg-disclaimer {
-  margin-top: 2px;
-  font-size: 12px;
-  color: var(--muted);
-}
-
-.controls {
-  display: flex;
-  gap: 12px;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-/* The status block owns spacing via gap; reset child component margins. */
-.lg-status :deep(.stats) {
-  margin: 0;
-}
-.lg-status :deep(.rank-indicator) {
-  margin: 0;
-}
-
-.range-form,
-.add-form {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-}
-
-.add-form input,
-.range-form input {
-  width: 160px;
-}
-
-.divider {
-  width: 1px;
-  height: 24px;
-  background: color-mix(in oklab, var(--primary) 16%, var(--surface));
-  margin: 0 8px;
-}
-
-.quick-actions button {
-  margin-right: 8px;
-}
-
-/* Recenter button floats inside chart-wrapper */
-.recenter-btn {
-  position: absolute;
-  bottom: 8px;
-  left: 8px;
-  z-index: 10;
-  width: 36px;
-  height: 36px;
-  padding: 0;
-  border-radius: 8px;
-  background: color-mix(in oklab, var(--surface) 85%, #000);
-  border: 1px solid color-mix(in oklab, var(--primary) 30%, var(--surface));
-  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.4);
-}
-.recenter-btn:hover {
-  background: color-mix(in oklab, var(--surface) 90%, #000);
-  box-shadow: 0 10px 26px rgba(0, 0, 0, 0.5), 0 0 0 3px var(--ring);
-}
-
-.quick-actions .spacer {
-  flex: 1;
-}
-
-.chart-actions {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  margin-top: 12px;
-}
-
-.chart-actions .action-group {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-.chart-actions .spacer {
-  flex: 1;
-}
-
-.btn-danger {
-  color: var(--danger);
-  border-color: color-mix(in oklab, var(--danger) 30%, var(--surface));
-}
-.btn-danger:hover {
-  border-color: color-mix(in oklab, var(--danger) 55%, var(--surface));
-  box-shadow: 0 8px 22px rgba(0, 0, 0, 0.45),
-    0 0 0 3px color-mix(in oklab, var(--danger) 30%, transparent);
-}
-
-.stats {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-  gap: 12px;
-  margin: 8px 0 12px;
-}
-
-.stat {
-  border: 1px solid color-mix(in oklab, var(--primary) 16%, var(--surface));
-  border-radius: 10px;
-  padding: 12px;
-  background: color-mix(in oklab, var(--surface) 85%, black);
-  box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.02),
-    0 6px 24px rgba(0, 0, 0, 0.35);
-}
-
-.stat-label {
-  font-size: 12px;
-  color: var(--muted);
-  margin-bottom: 4px;
-}
-
-.stat-value {
-  font-size: 24px;
-  font-weight: 800;
-  color: var(--text-strong);
-  text-shadow: 0 0 20px color-mix(in oklab, var(--primary) 35%, transparent);
-}
-
-
-.today-progress {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 6px 8px;
-  font-size: 13px;
-}
-.today-label {
-  color: var(--muted);
-  font-weight: 700;
-  text-transform: uppercase;
-  letter-spacing: 0.04em;
-  font-size: 12px;
-}
-.today-value {
-  color: var(--success);
-  font-weight: 800;
-}
-.today-value.today-loss {
-  color: var(--danger);
-}
-
-.status-banner {
-  padding: 10px 14px;
-  border-radius: 8px;
-  margin-bottom: 12px;
-  font-size: 13px;
-  font-weight: 600;
-}
-.auth-banner {
-  background: color-mix(in oklab, var(--warning, #f59e0b) 10%, var(--surface));
-  color: var(--muted);
-  border: 1px solid color-mix(in oklab, var(--warning, #f59e0b) 20%, var(--surface));
-}
-
-.chart-blurred {
-  filter: blur(4px);
-  opacity: 0.4;
-  transition: filter 0.3s, opacity 0.3s;
-  pointer-events: none;
-}
-
-.chart-overlay {
-  position: absolute;
-  inset: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  pointer-events: none;
-  z-index: 10;
-}
-
-.chart-overlay-text {
-  font-size: 15px;
-  font-weight: 600;
-  color: var(--text);
-  background: color-mix(in oklab, var(--surface) 80%, transparent);
-  padding: 10px 20px;
-  border-radius: 8px;
-}
-
-.chart-overlay-error {
-  color: #ff6b6b;
-}
-
-.error {
-  color: #ff6b6b;
-  margin: 4px 0 8px;
-}
-
-.chart-wrapper {
-  border: 1px solid color-mix(in oklab, var(--primary) 16%, var(--surface));
-  border-radius: 10px;
-  padding: 8px;
-  background: radial-gradient(
-      100% 100% at 0% 0%,
-      rgba(255, 212, 0, 0.05) 0%,
-      rgba(0, 0, 0, 0) 40%
-    ),
-    var(--surface);
-  position: relative;
-}
-
-svg {
-  width: 100%;
-  height: auto;
-  /* Enable custom gestures */
-  touch-action: none;
-  cursor: default;
-}
-
-svg.nav-disabled {
-  touch-action: auto;
-}
-
-.interaction-overlay {
-  fill: transparent;
-  cursor: grab;
-}
-.interaction-overlay:active {
-  cursor: grabbing;
-}
-
-.axes line {
-  stroke: color-mix(in oklab, var(--primary) 35%, #1f2937);
-  stroke-width: 1.25;
-}
-
-.rank-label {
-  font-size: 9px;
-  font-weight: 600;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  pointer-events: none;
-}
-
-.grid line {
-  stroke: color-mix(in oklab, var(--primary) 14%, #374151);
-  stroke-opacity: 0.5;
-  stroke-width: 1;
-}
-
-.line {
-  fill: none;
-  stroke: var(--accent);
-  stroke-width: 2.5;
-  filter: drop-shadow(
-    0 0 6px color-mix(in oklab, var(--accent) 60%, transparent)
-  );
-}
-
-.placement-baseline {
-  fill: none;
-  stroke: color-mix(in oklab, var(--accent) 55%, #9ca3af);
-  stroke-width: 2;
-  stroke-dasharray: 4 5;
-  opacity: 0.55;
-}
-
-.proj {
-  fill: none;
-  stroke: color-mix(in oklab, var(--primary) 45%, #9ca3af);
-  stroke-width: 2;
-  stroke-dasharray: 6 6;
-}
-
-.proj-from-last {
-  stroke: var(--success);
-}
-
-.deviation-wedge {
-  fill: var(--accent);
-  opacity: 0.08;
-  stroke: none;
-}
-
-.proj-average-pace {
-  stroke: var(--accent);
-  stroke-dasharray: 4 4;
-  opacity: 0.5;
-}
-
-.point circle {
-  fill: var(--danger);
-  filter: drop-shadow(0 0 4px rgba(255, 61, 127, 0.65));
-}
-
-/* Pace graph */
-.pace-axis {
-  stroke: color-mix(in oklab, var(--primary) 35%, #1f2937);
-  stroke-width: 1.25;
-}
-
-.pace-zero-line {
-  stroke: color-mix(in oklab, var(--primary) 30%, #6b7280);
-  stroke-width: 1;
-  stroke-dasharray: 4 4;
-}
-
-.pace-zero-label {
-  fill: color-mix(in oklab, var(--primary) 45%, #9ca3af);
-  font-size: 8px;
-  font-weight: 600;
-  pointer-events: none;
-}
-
-.pace-line {
-  fill: none;
-  stroke-width: 2;
-}
-
-.pace-line-required {
-  stroke: color-mix(in oklab, var(--primary) 45%, #9ca3af);
-  stroke-dasharray: 5 5;
-}
-
-.pace-line-earned {
-  stroke: var(--accent);
-  filter: drop-shadow(0 0 4px color-mix(in oklab, var(--accent) 50%, transparent));
-}
-
-.pace-point-required circle {
-  fill: color-mix(in oklab, var(--primary) 55%, #9ca3af);
-}
-
-.pace-point-earned circle {
-  fill: var(--accent);
-  filter: drop-shadow(0 0 3px color-mix(in oklab, var(--accent) 50%, transparent));
-}
-
-.pace-legend {
-  font-size: 8px;
-  font-weight: 600;
-  letter-spacing: 0.04em;
-}
-
-.pace-legend-required {
-  fill: color-mix(in oklab, var(--primary) 55%, #9ca3af);
-}
-
-.pace-legend-earned {
-  fill: var(--accent);
-}
-
-.labels text {
-  font-size: 12px;
-  fill: var(--muted);
-}
-
-.axis-label {
-  font-size: 11px;
-  fill: var(--muted);
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-  font-weight: 700;
-}
-
-/* Positioned version for inside the graph wrapper */
-.rank-in-graph {
-  position: absolute;
-  top: 8px;
-  left: 8px;
-  max-width: min(460px, 60%);
-  z-index: 9;
-}
-
-/* ── Mobile (phone) ── */
-@media (max-width: 480px) {
-  .chart-actions {
-    flex-wrap: wrap;
-  }
-  .chart-actions .action-group {
-    flex: 1 1 auto;
-  }
-  .chart-actions button {
-    flex: 1 1 auto;
-    min-width: 0;
-  }
-  .chart-actions .spacer {
-    display: none;
-  }
-
-  .controls {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .status-banner {
-    font-size: 12px;
-    padding: 8px 12px;
-  }
-
-  .lg-header {
-    padding: 0 4px;
-  }
-}
-</style>
+<style scoped src="./LineGraph.css"></style>
